Address PR comments regarding sparql expressions
