+++ src/engine/sparqlExpressions/SparqlExpressionValueGetters.cpp
@@ -94,7 +94,7 @@ std::optional<std::string> StringValueGetter::operator()(
 std::optional<std::string> ReplacementStringGetter::operator()(
     Id id, const EvaluationContext* context) const {
   std::optional<std::string> originalString =
-      LiteralFromIdGetter::operator()(id, context);
+      LiteralFromIdGetter{}(id, context);
   if (!originalString.has_value()) {
     return originalString;
   }
