+++ src/engine/sparqlExpressions/SparqlExpressionValueGetters.h
@@ -255,8 +255,7 @@ struct LiteralFromIdGetter : Mixin<LiteralFromIdGetter> {
 // they can be used by re2 as replacement strings. So '$1 \abc \$' becomes
 // '\1 \\abc $', where the former variant is valid in the SPARQL standard and
 // the latter represents the format that re2 expects.
-struct ReplacementStringGetter : LiteralFromIdGetter,
-                                 Mixin<ReplacementStringGetter> {
+struct ReplacementStringGetter : Mixin<ReplacementStringGetter> {
   using Mixin<ReplacementStringGetter>::operator();
   std::optional<std::string> operator()(ValueId,
                                         const EvaluationContext*) const;
